% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format}
\alias{format}
\alias{text_format}
\alias{file_format}
\alias{formatting_options}
\title{Format a JSON file or string}
\usage{
text_format(text, ..., formatting_options = NULL)

file_format(file, ..., formatting_options = NULL)

formatting_options(
  tab_size = 4L,
  insert_spaces = TRUE,
  eol = "\\n",
  insert_final_newline = TRUE
)
}
\arguments{
\item{text}{A single string containing JSON.}

\item{...}{These dots are for future extensions and must be empty.}

\item{formatting_options}{The result of \code{\link[=formatting_options]{formatting_options()}}. If
\code{NULL}, a default set of options are used.}

\item{file}{Path to file on disk. File must exist.}

\item{tab_size}{The number of spaces to use to indicate a single tab when
\code{insert_spaces} is \code{TRUE}.}

\item{insert_spaces}{Whether to use spaces or tabs for indentation.}

\item{eol}{The character used for the end of a line. This is only applicable
when the text doesn't already contain an existing line ending, i.e. an
empty string or a string spanning a single line.}

\item{insert_final_newline}{Whether or not to insert a final newline.}
}
\description{
Format a JSON file or string, preserving comments.
}
\examples{
text <- '{"foo":[1,2]}'
cat(text_format(text))

formatting_options <- formatting_options(tab_size = 2)
cat(text_format(text, formatting_options = formatting_options))
}
