% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit-package.R
\docType{package}
\name{jsonedit-package}
\alias{jsonedit}
\alias{jsonedit-package}
\title{jsonedit: JSON Parsing and Modification Utilities}
\description{
Bindings to 'node-jsonc-parser' to parse, format, and modify JSON files while retaining comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/jsonedit}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroenooms@gmail.com}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co}
}

Other contributors:
\itemize{
  \item Microsoft Corporation (node-jsonc-parser library) [copyright holder]
}

}
\keyword{internal}
