% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit-package.R
\docType{package}
\name{jsonedit-package}
\alias{jsonedit}
\alias{jsonedit-package}
\title{jsonedit: Whitespace Preserving JSON Editor}
\description{
Bindings to `node-jsonc-parser` to modify or format JSON files while retaining comments and whitespace.
}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroenooms@gmail.com}

}
\keyword{internal}
