% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit-package.R
\docType{package}
\name{jsonedit-package}
\alias{jsonedit}
\alias{jsonedit-package}
\title{jsonedit: JSON Parsing and Modification Utilities}
\description{
Bindings to 'node-jsonc-parser' to parse, format, and modify JSON files while retaining comments.
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroenooms@gmail.com}
}

Other contributors:
\itemize{
  \item Microsoft Corporation (node-jsonc-parser library) [copyright holder]
}

}
\keyword{internal}
